requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
crawlbase==1.0.0
idna==3.10
lxml==5.3.0
numpy==2.1.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
six==1.16.0
soupsieve==2.6
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
wikipedia==1.4.0
